Added classes game scenes
