+++ Data/Game/MatrixShips.js
@@ -1,7 +1,8 @@
 "use strict";
 import MessageBox from "../Modules/Blocks/MessageBox/MessageBox.js";
 
-import {createSecoundGameScene} from "./CreateGame.js";
+import FirstGameScene from "./GameSceneFirst.js";
+import SecondGameScene from "./GameSceneSecond.js";
 import ShipList from "./ShipList.js";
 
 function getMatrixShips () {
@@ -11,9 +12,10 @@ function getMatrixShips () {
     if (shipList.canDoMatrix()) {
         let matrixShips = shipList.CreateMatrix();
 
-        let allGame = document.getElementsByClassName("all_game");
-        document.body.removeChild(allGame[0]);
-        createSecoundGameScene(matrixShips);
+        let firstGameScene = new FirstGameScene();
+        firstGameScene.hide();
+        let secondGameScene = new SecondGameScene();
+        secondGameScene.show(matrixShips);
     }
     else {
         let mb = new MessageBox("Placement error", "You haven't placed all the ships!");
